+++ app/assets/config/manifest.js
@@ -4,4 +4,4 @@
 //= link_tree ../../../vendor/javascript .js
 //= link application.js
 //= link application.css
-
+//= link_tree ../images
