+++ app/assets/stylesheets/menu-bar.scss
@@ -1,0 +1,20 @@
+/* app/assets/stylesheets/menu-bar.scss */
+
+/* メニューバー */
+.menu-bar {
+  position: fixed;
+  top: 0;
+  left: -250px; /* 初期は非表示（左に隠す） */
+  width: 250px;
+  height: 100vh;
+  background-color: #9b80eb;
+  box-shadow: 2px 0 5px rgba(0, 0, 0, 0.2);
+  overflow-y: auto;
+  transition: left 0.3s ease; /* スライドのアニメーション */
+  z-index: 1000;
+}
+
+/* メニュー開いた時のスタイル */
+.menu-bar.open {
+  left: 0; /* メニュー表示 */
+}
