+++ app/controllers/application_controller.rb
@@ -6,9 +6,4 @@ class ApplicationController < ActionController::Base
   def current_user
     @current_user ||= User.find_by(id: session[:user_id])
   end
-
-  def test_log
-    Rails.logger.debug "ここはデバッグログのテストです！"
-    render plain: "ログ確認用のテストページ"
-  end
 end
