+++ app/views/shared/_header.html.erb
@@ -1,6 +1,7 @@
+<!-- app/views/shared/_header.html.erb -->
 <header class="header">
   <div class="header__left">
-    <button class="menu-button">☰</button>
+    <button class="menu-button" id="menu-toggle">☰</button>
   </div>
   <div class="header__right">
     <% if current_user %>
