+++ app/views/shared/_menu_bar.html.erb
@@ -1,4 +1,5 @@
-<nav class="menu-bar">
+<!-- app/views/shared/_menu_bar.html.erb -->
+<nav class="menu-bar" id="menu-bar">
   <ul class="menu-list">
     <li class="menu-item has-submenu">
       <span>ALPIとチャット</span>
@@ -8,26 +9,20 @@
         <li><%= link_to "宇宙人と話す", "#" %></li>
       </ul>
     </li>
-
     <li class="menu-item has-submenu">
       <span>山情報を探す</span>
       <ul class="submenu">
         <li><%= link_to "山情報一覧・検索", mountains_path %></li>
       </ul>
     </li>
-
     <li class="menu-item has-submenu">
       <span>レポートを見る</span>
       <ul class="submenu">
         <li><%= link_to "レポート一覧・検索", "#" %></li>
       </ul>
     </li>
-
     <li class="menu-item">
       <%= link_to "新規登録", "#" %>
-    </li>
-    <li class="menu-item">
-      <%= link_to "ログイン", "#" %>
     </li>
   </ul>
 </nav>
