+++ config/environments/development.rb
@@ -79,4 +79,5 @@ Rails.application.configure do
   # Apply autocorrection by RuboCop to files generated by `bin/rails generate`.
   # config.generators.apply_rubocop_autocorrect_after_generate!
   config.log_level = :debug
+  config.file_watcher = ActiveSupport::FileUpdateChecker
 end
