+++ docker-compose.dev.yml
@@ -1,4 +1,3 @@
-version: "3.9"
 services:
   db:
     image: postgres:latest
@@ -21,7 +20,9 @@ services:
     tty: true
     stdin_open: true
     volumes:
-      - .:/app
+      - type: bind
+        source: ./
+        target: /app
     environment:
       TZ: Asia/Tokyo
       DATABASE_HOST: db
@@ -35,3 +36,4 @@ services:
 
 volumes:
   db_data:
+  node_modules:
