+++ docker-compose.yml
@@ -19,7 +19,8 @@ services:
     tty: true
     stdin_open: true
     volumes:
-      - .:/alpi
+      - db_data:/var/lib/postgresql/data
+      - node_modules:/app/node_modules
     environment:
       TZ: Asia/Tokyo
       DATABASE_HOST: db
@@ -34,4 +35,6 @@ services:
       - db
 volumes:
   db_data:
+  # bundle_data
+  node_modules:
